Optimize lodash imports
Since only one function is used, it is better to import it separately
instead of the entire library. While it can be done manually, in the
future if needed, we might have to do it for every function we end up
using. Hence, the plugin is a more convenient way.